Extract SpigotUpdateRunnable to avoid NoClassDefFoundError on update check
